Refactor signature of bind
This commit change the trait bound on bind to FnMut to make it actually
works on practice. It has to call the mem::forget or the closure will
just be dropped.

The instance is also changed with Arc to make user able to reference
count them safely across thread (on Rust side, we "trust" original
webview knows about thread safety)

See example to get a look how these methods are called.